Fix ut_adm_license
Despite the fact that the license is free on github, it is still
necessary to have a correct license to be able to start.
The license contained in the code was outdated and failing.

This patch puts a correct test license (actually, the one used on
github) and fixes the unt test.