Allocate the ndev when vrt_open and release it when passing device down.
I aim to be able to make the import on open, and unimport on close, but
this is impossible for now because the size of the device is not know on
open.
What I would like to do is having some kind of import protocol between
the client and server, so that the client can get the information
directly from the server (just in the way iSCSI does) but for now, this
is not possible because the 'protocol' between client and server is
still a real crap and is very hard to rework...
Anyway, I am still working on it.

Signed-off-by: Sébastien Gonzalve <sebastien.gonzalve@aliceadsl.fr>