allow to use os_select instead of exa_select_in/out
Exanodes has a custom implementation of select() that relies on the
exa_common kernel module. This was historically implemented to fix
deadlocks occuring under heavy memory pressure when testing on Linux.

Nevertheless, this implementation makes it impossible for exanodes to
run without kernel modules, which may be painful for debugging or
testing.

This patch offers the ability to compile exanodes with the "standard"
Linux select().

Signed-off-by: Sébastien Gonzalve <sebastien.gonzalve@aliceadsl.fr>