remove exa_socket_tweak_emergency_pool()
This overkill stuff is here to handle some failing test on huge clusters
(over 64 nodes). The hack is modifying the server config 'under the
hood' and tries to restore it afterward.
Finaly, it also make the exa_common to be necessary in order to make
sure things done "under the hood" are undone "under the hood" thanks to
kernel notification of terminating processes.

I considere that this kind of hack deserve to be documented but not to
be build in. An intgrator may increase the emergency pool by modifying
the init script, and take the responsability for it.

This patch removes all this overkill dead stuff.

Signed-off-by: Sébastien Gonzalve <sebastien.gonzalve@aliceadsl.fr>