Rework rdev_test so that it become usable
This tool was almost useless as it was not able to take input not to
print out what it reads.

This patch makes the tool to get data from the standard input and print
it to the stndard output. This makes it easier to test the rdev layer by
itself, and especially when trying to integrate libaio.

Signed-off-by: OznOg <sebastien.gonzalve@aliceadsl.fr>