Complete rework of dkms related files
The legacy code was using some hugly tricks to allow to build or not
some modules.
The problem is that this was of doing was specially obscure, poorly
scalable and complicated.

This patch changes the makefile so that we do not need to comment out
rules, but just not call them, and rework the dkms.conf so that it
constructs dynamically the array of modules to build, this will provide
ability do scale.