Cleanup extra kernel flags
For some reason, many flags were passed to kernel when they are actually
useless (the kernel code does not use them).
This was creating confusion and hiding the mandatory flags to be passed:
When activating perfs, exa_bd structures change, thus kernel MUST know
that perfs are on or of in order to align data correctly.

This patch performs a cleanup that remove all useless flags and keep
only the mandatory one.

Signed-off-by: OznOg <sebastien.gonzalve@aliceadsl.fr>