Relax unit test on socket timeout
The timeout set on socket may be round by the kernel, thus the value we get
back when calling get, may be different. For some reason it happen on Travis
and not on my local machines; whatever, the test was too strict.

This patch just checks that the kernel kept a value that is at least bigger
than the one we passed; This should fix Travis checks.