Fix back to async the connection to nodes with CLI
This code was born asynchronous (non blocking connect() call) but was
put synchronous while porting on windows as it seems so painful to make
it work on windows.

That said, who cares of windows? If people want to use windows, the get
the slow synchronous version, if they use Linux, they benefit from the
async version.

People are already used to windows sloooooooooowness after all.